requirements.txt
Django==4.2.1
django-allauth==0.54.0
django-cors-headers==4.0.0
django-environ==0.10.0
djangorestframework==3.14.0
numpy==1.24.3
pandas==2.0.1
Pillow==9.5.0
psycopg2-binary==2.9.6
PyJWT==2.7.0
requests==2.31.0
selenium==4.10.0
langchain
openai
streamlit
django-extensions==3.2.3
django-redis==5.3.0